Move taskmanager to task package
